+++ Pinny Notes/Enums/TransparencyModes.cs
@@ -1,0 +1,8 @@
+﻿namespace PinnyNotes.WpfUi.Enums;
+
+public enum TransparencyModes
+{
+    Disabled = 0,
+    Enabled = 1,
+    OnlyWhenPinned = 2
+}